# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Techno Terminal API",
    "description": "RESTful backend for the Techno Terminal CRM & Operations System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/attendance/session/{session_id}": {
      "get": {
        "tags": [
          "Attendance",
          "Attendance"
        ],
        "summary": "Get session roster with attendance status",
        "description": "Returns all enrolled students for this session with their current attendance status.",
        "operationId": "get_session_attendance_api_v1_attendance_session__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_SessionAttendanceRowDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/session/{session_id}/mark": {
      "post": {
        "tags": [
          "Attendance",
          "Attendance"
        ],
        "summary": "Mark / update attendance for a session",
        "description": "Bulk upsert attendance for a session. Pass a dict of student_id \u2192 status. Students not in the dict are left unchanged (not set to absent).",
        "operationId": "mark_attendance_api_v1_attendance_session__session_id__mark_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkAttendanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MarkAttendanceResponseDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/finance/revenue-by-date": {
      "get": {
        "tags": [
          "Analytics \u2014 Financial",
          "Analytics \u2014 Financial"
        ],
        "summary": "Get revenue breakdown by date",
        "description": "Returns daily revenue totals within the specified date range.",
        "operationId": "get_revenue_by_date_api_v1_analytics_finance_revenue_by_date_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start date (YYYY-MM-DD)",
              "title": "Start"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End date (YYYY-MM-DD)",
              "title": "End"
            },
            "description": "End date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_RevenueByDateDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/finance/revenue-by-method": {
      "get": {
        "tags": [
          "Analytics \u2014 Financial",
          "Analytics \u2014 Financial"
        ],
        "summary": "Get revenue breakdown by payment method",
        "description": "Returns revenue totals grouped by payment method within the date range.",
        "operationId": "get_revenue_by_method_api_v1_analytics_finance_revenue_by_method_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start date (YYYY-MM-DD)",
              "title": "Start"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End date (YYYY-MM-DD)",
              "title": "End"
            },
            "description": "End date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_RevenueByMethodDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/finance/outstanding-by-group": {
      "get": {
        "tags": [
          "Analytics \u2014 Financial",
          "Analytics \u2014 Financial"
        ],
        "summary": "Get outstanding balances by group",
        "description": "Returns outstanding balance totals grouped by academic group.",
        "operationId": "get_outstanding_by_group_api_v1_analytics_finance_outstanding_by_group_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_OutstandingByGroupDTO__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/finance/top-debtors": {
      "get": {
        "tags": [
          "Analytics \u2014 Financial",
          "Analytics \u2014 Financial"
        ],
        "summary": "Get top debtors",
        "description": "Returns students with highest outstanding balances.",
        "operationId": "get_top_debtors_api_v1_analytics_finance_top_debtors_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of results to return",
              "default": 15,
              "title": "Limit"
            },
            "description": "Number of results to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_TopDebtorDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_MarkAttendanceResponseDTO_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarkAttendanceResponseDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[MarkAttendanceResponseDTO]"
      },
...
      "ApiResponse_list_OutstandingByGroupDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OutstandingByGroupDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[OutstandingByGroupDTO]]"
      },
...
      "ApiResponse_list_RevenueByDateDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RevenueByDateDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[RevenueByDateDTO]]"
      },
      "ApiResponse_list_RevenueByMethodDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RevenueByMethodDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[RevenueByMethodDTO]]"
      },
...
      "ApiResponse_list_SessionAttendanceRowDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SessionAttendanceRowDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[SessionAttendanceRowDTO]]"
      },
...
      "ApiResponse_list_TopDebtorDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TopDebtorDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[TopDebtorDTO]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarkAttendanceRequest": {
        "properties": {
          "student_statuses": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Student Statuses"
          }
        },
        "type": "object",
        "required": [
          "student_statuses"
        ],
        "title": "MarkAttendanceRequest",
        "description": "Body for POST /attendance/session/{session_id}/mark.\nstudent_statuses maps student_id \u2192 status string (\"present\" | \"absent\" | \"late\" | \"excused\")."
      },
      "MarkAttendanceResponseDTO": {
        "properties": {
          "marked": {
            "type": "integer",
            "title": "Marked"
          },
          "skipped": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "marked",
          "skipped"
        ],
        "title": "MarkAttendanceResponseDTO",
        "description": "Output DTO detailing how many records were affected."
      },
      "OutstandingByGroupDTO": {
        "properties": {
          "group_id": {
            "type": "integer",
            "title": "Group Id"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "students_with_balance": {
            "type": "integer",
            "title": "Students With Balance"
          },
          "total_outstanding": {
            "type": "number",
            "title": "Total Outstanding"
          }
        },
        "type": "object",
        "required": [
          "group_id",
          "group_name",
          "course_name",
          "students_with_balance",
          "total_outstanding"
        ],
        "title": "OutstandingByGroupDTO"
      },
...
      "RevenueByDateDTO": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "net_revenue": {
            "type": "number",
            "title": "Net Revenue"
          }
        },
        "type": "object",
        "required": [
          "day",
          "net_revenue"
        ],
        "title": "RevenueByDateDTO"
      },
      "RevenueByMethodDTO": {
        "properties": {
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "net_revenue": {
            "type": "number",
            "title": "Net Revenue"
          },
          "receipt_count": {
            "type": "integer",
            "title": "Receipt Count"
          }
        },
        "type": "object",
        "required": [
          "payment_method",
          "net_revenue",
          "receipt_count"
        ],
        "title": "RevenueByMethodDTO"
      },
...
      "SessionAttendanceRowDTO": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "present",
              "absent",
              "late",
              "excused"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "status"
        ],
        "title": "SessionAttendanceRowDTO",
        "description": "Output DTO for a student's attendance in a single session."
      },
...
      "TopDebtorDTO": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "parent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Name"
          },
          "phone_primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Primary"
          },
          "total_outstanding": {
            "type": "number",
            "title": "Total Outstanding"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "student_name",
          "parent_name",
          "phone_primary",
          "total_outstanding"
        ],
        "title": "TopDebtorDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}